backend/requirements.txt
# Backend Dependencies for Virtual Mirror API

# Core FastAPI dependencies
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12
pydantic==2.9.2

# Environment variables
python-dotenv==1.0.0

# CORS support
fastapi-cors==0.0.6

# PDF generation
reportlab==4.2.5

# Optional: Google Text-to-Speech
gtts==2.5.3

# Data processing
numpy==1.26.4
pandas==2.2.3

# Additional utilities
python-dateutil==2.9.0

# Database (PostgreSQL/Supabase) - Async support
sqlalchemy[asyncio]==2.0.35
asyncpg==0.29.0
alembic==1.13.3
greenlet==3.1.1
